URLBuffer is an interface with a simple implementation + implementations are pluggable using config
